# openapi.yaml
openapi: 3.1.0
info:
  title: Dropless API
  version: "0.1.0"
  description: |
    Self-hosted, zero-loss webhook delivery. A message is only acknowledged with
    `201` after its event and fan-out deliveries are committed to Postgres in a
    single transaction. Delivery is at-least-once (effectively-once) with an
    immutable `Idempotency-Key` per delivery.

    All `/v1/*` routes require `Authorization: Bearer <api-key>`. Outbound
    webhooks are signed Svix-style (`webhook-id`, `webhook-timestamp`,
    `webhook-signature: v1,<base64(HMAC_SHA256(secret, "{id}.{ts}.{body}"))>`).
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /v1/endpoints:
    post:
      tags: [Endpoints]
      summary: Register a subscriber endpoint
      operationId: createEndpoint
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEndpoint" }
      responses:
        "201":
          description: Created. The signing `secret` is returned ONCE.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedEndpoint" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    get:
      tags: [Endpoints]
      summary: List the tenant's endpoints
      operationId: listEndpoints
      responses:
        "200":
          description: Endpoints (secrets never included).
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Endpoint" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /ingest/{slug}:
    post:
      tags: [Inbound]
      summary: Receive a provider webhook (raw-first, deduped, bridged to outbound)
      operationId: ingest
      description: |
        Authenticated by the provider's signature (Stripe / GitHub / HMAC), not
        an API key. The raw body is persisted before anything else, deduped on
        the provider's event id, and bridged into an outbound event.
      security: []
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json: { schema: { type: object, additionalProperties: true } }
      responses:
        "202": { description: Accepted, persisted, and bridged. }
        "200": { description: Duplicate (already received) — a no-op. }
        "401": { description: Signature verification failed. }
        "404": { description: Unknown or disabled source. }
  /v1/inbound-sources:
    post:
      tags: [Inbound]
      summary: Register an inbound source
      operationId: createInboundSource
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/CreateInboundSource" } }
      responses:
        "201":
          description: Created. The `secret` is returned ONCE.
          content:
            application/json: { schema: { $ref: "#/components/schemas/CreatedInboundSource" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    get:
      tags: [Inbound]
      summary: List the tenant's inbound sources
      operationId: listInboundSources
      responses:
        "200":
          description: Inbound sources (secrets never included).
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/InboundSource" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, description: An API key from `dropless create-api-key`. }
...
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    BadRequest:
      description: Invalid request body.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      properties: { error: { type: string } }
      required: [error]
...
    Endpoint:
      type: object
      required: [id, tenant_id, url, disabled, circuit_state, consecutive_failures, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string }
        url: { type: string, format: uri }
        disabled: { type: boolean }
        circuit_state: { type: string, enum: [closed, open, half_open] }
        circuit_open_until: { type: [string, "null"], format: date-time }
        consecutive_failures: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateEndpoint:
      type: object
      required: [url]
      properties:
        url: { type: string, format: uri }
        secret: { type: string, description: "Optional; a whsec_… secret is generated if omitted." }
    CreatedEndpoint:
      allOf:
        - { $ref: "#/components/schemas/Endpoint" }
        - type: object
          required: [secret]
          properties:
            secret: { type: string, description: "Signing secret — returned ONLY here." }
...
    InboundSource:
      type: object
      required: [id, tenant_id, slug, provider, disabled, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string }
        slug: { type: string }
        provider: { type: string, enum: [stripe, github, hmac] }
        disabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateInboundSource:
      type: object
      required: [slug, provider]
      properties:
        slug: { type: string, description: "Globally-unique URL segment for POST /ingest/{slug}." }
        provider: { type: string, enum: [stripe, github, hmac] }
        secret: { type: string, description: "Provider signing secret; generated for hmac if omitted." }
    CreatedInboundSource:
      allOf:
        - { $ref: "#/components/schemas/InboundSource" }
        - type: object
          required: [secret]
          properties:
            secret: { type: string, description: "Verification secret — returned ONLY here." }